+++ src/main/java/entidad/Especialidad.java
@@ -28,7 +28,7 @@ public class Especialidad implements Serializable{
 	@Id
 	@Column(name="id")
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 	
 	@Column(name="nombre",unique=true)
 	private String nombre;
@@ -36,7 +36,7 @@ public class Especialidad implements Serializable{
 	// Constructor vacio requerido por Hibernate
 	public Especialidad() {}
 	
-	public Especialidad(Long id, String nombre) {
+	public Especialidad(int id, String nombre) {
 		this.id = id;
 		this.nombre = nombre;
 	}
@@ -46,11 +46,11 @@ public class Especialidad implements Serializable{
 	}
 
 	// Getters y Setters
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
